requirements.txt
streamlit
Pillow
tweepy
